requirements.txt
openai
python-dotenv
langchain
langchain_community
langchain-openai
